+++ static_precompiler/compilers/__init__.py
@@ -3,3 +3,4 @@ from static_precompiler.compilers.coffeescript import CoffeeScript
 from static_precompiler.compilers.babel import Babel
 from static_precompiler.compilers.scss import SASS, SCSS
 from static_precompiler.compilers.less import LESS
+from static_precompiler.compilers.stylus import Stylus